Add a basic top file and a state
A `top.sls` file dictates which states will be applied to which minions.
Each state is represented in a dedicated .sls file that has the same
name as a state name listed in a `top.sls` file.

A `nodejs-package.sls` file is a typical exaple of a state file. It has
a package name (`nodejs`), then a command `pkg` that instructs an OS to
treat the above name as a package, and finally an argument `installed`
is passed to a `pkg` module saying it to be installed.