Add new state with multiple state modules
In a `hwaas-site` state I defined multiple state modules. One is to
create new user in a system with module `user.present`. Another is to
install a new package in a system. I can give an arbitrary name to this
sub-state, but I have to provide a valid value to a `name` key. And a
third one is to download the source code with a `git.latest` module.