Install package from a custom repository
Packages from custom repository are installed with `pkgrepo.managed`
state module, that defines first a repository. Next, should be a
sub-state that will install that customer package.

To run a command that depends on another sub-state, I used `cmd.wait`
state-module. A required key for this module is `watch` that specifies a
 sub-state that it depends on. If more than one sub-states are watching
 the same sub-state, that they are executed in the order how they are
 presented in a state file.