Add a dependency of an entire .sls file
It is possible to set up a dependency in a sub-state to an entire state
in an another .sls file. Use key `sls` under `require` and include the
filename also at the start of a current state file under `include`.

Here I'm asking `hwaas-source` sub-state to run only if all sub-states
defined in `nodejs-package` ran successfully.