Define a new top-level `hwaas-service` state
That state replaces `hwaas-site` in a `top.sls`, because the
`hwaas-service` is requiring `hwaas-service` as a dependency.

New `hwaas-service` state introduces a new module - `supervisor`. It is
a Linux level system that allows to manage services in Linux. It is
installed and is "watching" for any changes made a in configuration file
 that describes what command to run, which user to use and location
 where to run a command.