Add SSL state to a load-balancer
To enable SSL on load-balancer, I need to add two files on Nginx: a
certificate and a key for a certificate. First, I created a pillar where
I specified a location of both a certificate and a key. Then in the same
state I put the content of both the certificate and key.

Next, in `init.sls` for a load-balancer I added two more sub-states that
basically create two files (one for a certificate and one for a key).
Here a Jinja2 template is used. Sub-state name is an equivalent of a
path to a file that I want to create. Here I'm telling with Jinja that I
need to create a file under the path that is written in key `cert-path`
under `hwaas-ssl` state. Next, using a `contents_pillar` I'm populating
the file with the content that is writtent under the value passed to
`contents_pillar`.

I'm using Jinja2 template inside a configuration file for Nginx to
specify location of a certificate and a key. In order for .conf file to
understand Jinja2 template, I need to pass a `template` key with a
respective value under .conf sub-state.

Nginx should be restarted if SS certificate and key is changed. That's
why I add them under a `watch` key of a `nginx` substate.