Add configurations for salt master and minions
With Vagrant it is possible to add a configuration file for both master
and minions. Thus, it is not needed then to SSH into each minion
machine, install Salt and set up master's address.

Added also to master a configuration to auto accept minions' keys. That
is not a recommended configuration, but here I made it just for a
simplicity.